% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crs_longlat}
\alias{crs_longlat}
\title{Lat-long coordinate reference system}
\format{
CRS
}
\usage{
crs_longlat
}
\description{
Coordinate reference system (CRS) for creating maps from
longitude-latitude coordinates. Used in analysis, but
not recommended for plots.
}
\details{
\code{crs_longlat} is EPSG4326
}
\seealso{
\code{\link{crs_Atlantic}}, \code{\link{crs_Pacific}},
 \code{\link{crs_S}}, \code{\link{crs_N}}
}
\keyword{datasets}
