% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crs_N}
\alias{crs_N}
\title{Arctic-centred coordinate reference system}
\format{
CRS
}
\usage{
crs_N
}
\description{
Coordinate reference system (CRS) for plotting and analysing maps.
WGS 84 / Arctic Polar Stereographic. Used in analysis, but
not recommended for plots.
}
\details{
crs_N is EPSG3995
}
\seealso{
\code{\link{crs_Atlantic}}, \code{\link{crs_Pacific}},
 \code{\link{crs_120E}}, \code{\link{crs_longlat}}, \code{\link{crs_S}}
}
\keyword{datasets}
