% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crs_Pacific}
\alias{crs_Pacific}
\title{Pacific-centred coordinate reference system}
\format{
CRS
}
\usage{
crs_Pacific
}
\description{
Coordinate reference system (CRS) for plotting and analysing maps.
Pacific-centred.
}
\details{
"+proj=robin +lon_0=180 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m
+no_defs"
}
\seealso{
\code{\link{crs_Atlantic}}, \code{\link{crs_120E}},
  \code{\link{crs_N}}, \code{\link{crs_S}}
}
\keyword{datasets}
