% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crs_S}
\alias{crs_S}
\title{Antarctic-centred coordinate reference system}
\format{
CRS
}
\usage{
crs_S
}
\description{
Coordinate reference system (CRS) for plotting and analysing maps.
WGS 84 / Antarctic Polar Stereographic. Used in analysis, but
not recommended for plots.
}
\details{
crs_N is EPSG 3031
}
\seealso{
\code{\link{crs_Atlantic}}, \code{\link{crs_Pacific}},
 \code{\link{crs_120E}}, \code{\link{crs_longlat}}, \code{\link{crs_N}}
}
\keyword{datasets}
