% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himach-package.R
\docType{package}
\name{himach-package}
\alias{himach}
\alias{himach-package}
\title{himach: High Mach Finds Routes for Supersonic Aircraft}
\description{
For supersonic aircraft, flying subsonic over land, find the best route between airports. Allow for coastal buffer and potentially closed regions. Use a minimal model of aircraft performance: the focus is on time saved versus subsonic flight, rather than on vertical flight profile. For modelling and forecasting, not for planning your flight!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/david6marsh/himach}
  \item Report bugs at \url{https://github.com/david6marsh/himach/issues}
}

}
\author{
\strong{Maintainer}: David Marsh \email{david6marsh@gmail.com}

Other contributors:
\itemize{
  \item Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} [contributor]
  \item  EUROCONTROL [funder, copyright holder]
}

}
\keyword{internal}
