% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{hm_clean_cache}
\alias{hm_clean_cache}
\title{Clean the route and SID-STAR cache.}
\usage{
hm_clean_cache(cache = c("route", "star"))
}
\arguments{
\item{cache}{Which caches to clear. Default is both \code{c("route",
"star")}.}
}
\value{
TRUE silently
}
\description{
Empties the cache.
}
\examples{
hm_clean_cache("route")

hm_clean_cache()

}
\seealso{
For more details see the cache section in the vignette:
  \code{vignette("Supersonic_Routes_in_depth", package = "himach")}. or
  \href{../doc/Supersonic_Routes_in_depth.html#cache}{Vignette on caching}
}
