% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{hm_load_cache}
\alias{hm_load_cache}
\title{Load route and SID/STAR cache}
\usage{
hm_load_cache(file)
}
\arguments{
\item{file}{Including the path.}
}
\value{
Invisible true
}
\description{
This silently overwrites any existing values in the cache.
}
\examples{
# not run
# hm_load_cache(file="") #load from this file

}
\seealso{
For more details see the cache section in
   `vignette("Supersonic_Routes_in_depth")`.
}
