% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mach_kph}
\alias{mach_kph}
\title{Speed of sound, for Mach to km conversion}
\format{
double
}
\usage{
mach_kph
}
\description{
1 Mach is approximately 1062kph in standard met conditions at the altitude
for supersonic flight (approx 50,000 feet).
}
\keyword{datasets}
