% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZ_buffer30}
\alias{NZ_buffer30}
\title{New Zealand polygons with 30km land buffer}
\format{
MULTIPOLYGON projected into \code{\link{crs_Pacific}}
}
\source{
\url{https://datafinder.stats.govt.nz/layer/104266-territorial-authority-2020-clipped-generalised/}
}
\usage{
NZ_buffer30
}
\description{
A dataset containing sf::MULTIPOLYGONS for New Zealand. Uses simplified
version of Stats NZ data, at 1km resolution, as in \code{\link{NZ_coast}}, with added 30km
buffer to all coasts.
}
\seealso{
\code{\link{NZ_coast}}, \code{\link{NZ_Buller_buffer40}}
}
\keyword{datasets}
