% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZ_Buller_buffer40}
\alias{NZ_Buller_buffer40}
\title{New Zealand Buller district plus 40km buffer}
\format{
MULTIPOLYGON projected into \code{\link{crs_Pacific}}
}
\source{
\url{https://datafinder.stats.govt.nz/layer/104266-territorial-authority-2020-clipped-generalised/}
}
\usage{
NZ_Buller_buffer40
}
\description{
Chosen purely for convenience of examples, a subset of \code{\link{NZ_coast}} for
Buller District only to which a buffer of 40km has been added. This forces
deviation in routes, when used as an 'avoid' dataset.
}
\seealso{
\code{\link{NZ_coast}}
}
\keyword{datasets}
