% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZ_coast}
\alias{NZ_coast}
\title{Simplified polygons of New Zealand.}
\format{
MULTIPOLYGON projected into \code{\link{crs_Pacific}}
}
\source{
\url{https://datafinder.stats.govt.nz/layer/104266-territorial-authority-2020-clipped-generalised/}
}
\usage{
NZ_coast
}
\description{
A dataset containing sf::MULTIPOLYGONS for New Zealand. Simplified version of
Stats NZ data, at 1km resolution. Called 'coast' but contains a bit more detail.
}
\seealso{
\code{\link{NZ_buffer30}}
}
\keyword{datasets}
