% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZ_grid}
\alias{NZ_grid}
\title{30km routing grid around New Zealand}
\format{
\linkS4class{GridLat}
}
\usage{
NZ_grid
}
\description{
Latitude-longitude-based routing grid around New Zealand at 30km target
distance, as generated by \code{\link{make_route_grid}}.
}
\details{
Uses built-in map \code{\link{NZ_buffer30}} projected to
\code{\link{crs_Pacific}}.
}
\keyword{datasets}
