% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZ_routes}
\alias{NZ_routes}
\title{5 example routes around New Zealand}
\format{
dataframe with 49 fields.
}
\usage{
NZ_routes
}
\description{
Some very unlikely supersonic routes around New Zealand using the test
aircraft that was given a very short range and slow subsonic cruise to get
the example to 'work'. [Not for operational use!]
}
\details{
Includes one refuelling stop (!) in Wellington.

Generated by \code{\link{find_route}}. Uses built-in map
\code{\link{NZ_buffer30}}.
}
\keyword{datasets}
