% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{profile_routes}
\alias{profile_routes}
\title{Profile a set of routes}
\usage{
profile_routes(
  routes,
  yvar = c("hours", "longitude"),
  ap_loc = make_airports(warn = FALSE),
  n_max = 2
)
}
\arguments{
\item{routes}{as generated by \code{\link{find_route}}}

\item{yvar}{horizontal axis is hours or longitude}

\item{ap_loc}{Airports and coordinates, by (silent) default from \code{\link{make_airports}}}

\item{n_max}{maximum number of routes to plot (default 2)}
}
\value{
A list of named list pairs of plots, which can be displayed using eg \code{result[1]}.
}
\description{
Profile a set of routes
}
\examples{
# not run ---
# plot_list <- profile_routes(routes, n_max = 3)
# plot_list # to display them all

}
