% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{profile_routes}
\alias{profile_routes}
\title{Profile a set of routes}
\usage{
profile_routes(
  routes,
  xvar = c("hours", "longitude"),
  ap_loc = make_airports(warn = FALSE),
  n_max = 2
)
}
\arguments{
\item{routes}{as generated by \code{\link{find_route}}}

\item{xvar}{horizontal axis is hours or longitude}

\item{ap_loc}{Airports and coordinates, by (silent) default from \code{\link{make_airports}}}

\item{n_max}{maximum number of routes to plot (default 2)}
}
\value{

}
\description{
Profile a set of routes
}
